requirements.txt
beautifulsoup4==4.12.3
numpy==1.26.4
pandas==2.2.2
setuptools==75.1.0
tabulate==0.9.0
fpdf2>=2.8.1